#2112 Code cleanup: moving the classes in more user friendly packages
https://github.com/metasfresh/metasfresh/issues/2112
