+++ de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/empties/impl/HUEmptiesService.java
@@ -25,8 +25,8 @@ import de.metas.handlingunits.inout.IReturnsInOutProducer;
 import de.metas.handlingunits.model.I_DD_NetworkDistribution;
 import de.metas.handlingunits.model.I_M_InOutLine;
 import de.metas.handlingunits.model.I_M_Locator;
+import de.metas.handlingunits.spi.impl.HUPackingMaterialDocumentLineCandidate;
 import de.metas.inout.IInOutDAO;
-import de.metas.inoutcandidate.spi.impl.HUPackingMaterialDocumentLineCandidate;
 import de.metas.material.planning.ddorder.IDistributionNetworkDAO;
 
 /*
