+++ de.metas.handlingunits.base/src/test/java/de/metas/handlingunits/allocation/transfer/HUTransformServiceTests.java
@@ -33,7 +33,7 @@ import de.metas.handlingunits.model.I_M_HU;
 import de.metas.handlingunits.model.I_M_HU_PI_Item;
 import de.metas.handlingunits.model.I_M_HU_PI_Item_Product;
 import de.metas.handlingunits.model.X_M_HU;
-import de.metas.inoutcandidate.spi.impl.IHUPackingMaterialCollectorSource;
+import de.metas.handlingunits.spi.IHUPackingMaterialCollectorSource;
 import mockit.Mocked;
 
 /*
