+++ de.metas.handlingunits.base/src/test/java/de/metas/handlingunits/allocation/transfer/HUTransformTestsBase.java
@@ -34,8 +34,8 @@ import de.metas.handlingunits.model.I_M_Locator;
 import de.metas.handlingunits.model.X_M_HU;
 import de.metas.handlingunits.model.X_M_HU_Status;
 import de.metas.handlingunits.model.validator.M_HU;
+import de.metas.handlingunits.spi.IHUPackingMaterialCollectorSource;
 import de.metas.handlingunits.trace.HUTransformTracingTests;
-import de.metas.inoutcandidate.spi.impl.IHUPackingMaterialCollectorSource;
 import de.metas.interfaces.I_M_Warehouse;
